+++ package.json
@@ -21,7 +21,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.0",
-    "react-slick": "^0.30.2",
+    "react-slick": "^0.30.1",
     "react-swiper": "^0.1.5",
     "slick-carousel": "^1.8.1",
     "swiper": "^11.0.6",
