+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./Pages/Home";
-import About from "./Pages/About";
+import About from "./Pages/About"
 import Login from "./Components/Auth-Components/Login";
 import Registration from "./Components/Auth-Components/Registration";
 import ContactUs from "./Pages/ContactUs";
@@ -10,8 +10,8 @@ function App() {
   return (
     <div className="App">
       <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Home />} />
+        <Routes> 
+          <Route path="/Home" element={<Home />} />
           <Route path="/About" element={<About />} />
           <Route path="/sign-in" element={<Login />} />
           <Route path="/sign-out" element={<Registration />} />
