+++ src/Components/About-Component/About_Style.css
@@ -1,0 +1,52 @@
+body {
+  background: #000428;
+  font-family: 'Roboto';
+}
+
+.bar {
+  animation: grow-bar cubic-bezier(0.23, 1, 0.32, 1) forwards;
+  background: linear-gradient(to left, #004e92, #000428);
+  opacity: 0;
+}
+
+.name {
+  animation: slide-name ease forwards;
+  opacity: 0;
+}
+
+.steps {
+  animation: grow-steps ease forwards;
+  opacity: 0;
+}
+
+@keyframes grow-steps {
+  0% {
+    transform: scale(.25);
+  }
+  
+  100% {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
+
+@keyframes slide-name {
+  0% {
+    transform: translateX(-25px);
+  }
+  
+  100% {
+    opacity: 1;
+    transform: translateX(0px);
+  }
+}
+
+@keyframes grow-bar {
+  0% {
+    width: 0;
+  }
+  100% {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
