--- src/Components/About-Component/Hero.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-// import { motion } from "framer-motion";
-
-const Hero = () => {
-  //  animation variants for the image
-
-  return (
-    <div className="contact mt-16 w-full bg-cover bg-center bg-no-repeat ">
-      <div className="w-1/2 flex flex-col-reverse md:flex-row items-center justify-center md:justify-between md:h-[80vh] ">
-        <div className="left-container text-white w-full px-8 md:px-12 lg:px-20">
-          <div className="dancing-script text-4xl md:text-6xl text-center md:text-left">
-            <h1 className="text-center">About Us: </h1>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Hero;
